tests/wpt/tests/css/css-inline/baseline-source/baseline-source-computed.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#baseline-source">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target"></div>

<script>
test_computed_value('baseline-source', 'auto');
test_computed_value('baseline-source', 'first');
test_computed_value('baseline-source', 'last');
</script>
